// src/models/itemProjetConsulterConv.ts
export interface ItemProjetConsulterConvI {
    id: number;
    idStatus: number;
    status: string;
    txtColorStatus: string;
    bgColorStatus: string;
    titleProjet: string;
    coutGlobal: number;
    contribution: number;
    situationDeblocage: number;
    dateDerniereSituation: string;
}